+++ examples/blank/rspack-entries/dev-entry.js
@@ -1,14 +1,17 @@
 import { DevApp } from '@botonic/react'
-import { routes } from '../src/routes'
+
+import { config } from '../src'
 import { locales } from '../src/locales'
 import { plugins } from '../src/plugins'
+import { routes } from '../src/routes'
 import { webchat } from '../src/webchat'
-import { config } from '../src'
+import { webviews } from '../src/webviews'
 
 export const app = new DevApp({
   routes,
   locales,
   plugins,
+  webviews,
   ...webchat,
   ...config,
 })