+++ examples/tutorial/esbuild-entries/node-entry.js
@@ -1,8 +1,0 @@
-import { NodeApp } from '@botonic/react/lib/esm'
-
-import { config } from '../src'
-import { locales } from '../src/locales'
-import { plugins } from '../src/plugins'
-import { routes } from '../src/routes'
-
-module.exports = new NodeApp({ routes, locales, plugins, ...config })