+++ packages/botonic-plugin-flow-builder/src/action/first-interaction.ts
@@ -10,11 +10,35 @@ import { getContentsByPayload } from './payload'
 export async function getContentsByFirstInteraction(
   context: FlowBuilderContext
 ): Promise<FlowContent[]> {
-  const { flowBuilderPlugin, request } = context
+  const { contentID, flowBuilderPlugin, request } = context
+
+  /*
+   * If the contentID is provided, the firstInteractionContents are obtained even if they are not used
+   * because when obtain this firstInteractionContents is when the session.flow_thread_id is updated.
+   * This is needed for example when send a WhatsApp campaign is sent,
+   * the bot not receives the message because this message is sent directly by the backend
+   * we expect the bot to respond only with the contents of the contentID and not with the firstInteractionContents.
+   */
   const firstInteractionContents = await flowBuilderPlugin.getStartContents()
 
-  // If the first interaction has a FlowBotAction, it should be the last content
-  // and avoid to render the match with keywords,intents or knowledge base
+  if (contentID) {
+    try {
+      const contentsByContentID =
+        await flowBuilderPlugin.getContentsByContentID(contentID)
+
+      if (contentsByContentID.length > 0) {
+        return contentsByContentID
+      }
+    } catch (error) {
+      console.warn(
+        `The contentID ${contentID} is not found. Returning the firstInteractionContents`
+      )
+    }
+  }
+
+  /* If the first interaction has a FlowBotAction, it should be the last content
+   * and avoid to render the match with keywords,intents or knowledge base
+   */
   if (firstInteractionContents.at(-1) instanceof FlowBotAction) {
     return firstInteractionContents
   }