+++ packages/botonic-plugin-flow-builder/src/types.ts
@@ -7,7 +7,10 @@ import {
 } from '@botonic/core'
 
 import { FlowContent } from './content-fields'
-import { HtFlowBuilderData } from './content-fields/hubtype-fields'
+import {
+  HtFlowBuilderData,
+  HtRatingButton,
+} from './content-fields/hubtype-fields'
 
 export interface InShadowingConfig {
   allowKeywords: boolean
@@ -111,4 +114,9 @@ export interface SmartIntentResponse {
 
 export interface PayloadParamsBase {
   followUpContentID?: string
+}
+
+export interface RatingSubmittedInfo extends HtRatingButton {
+  possibleOptions: string[]
+  possibleValues: number[]
 }